Added can_mag files, minor changes to mag.h and mag.c to convert floats to int output

+++ SAMM/SAMM_Mag/Inc/mag.h
@@ -24,7 +24,7 @@ uint16_t mag_write(mag *mag_dev, uint8_t reg, uint16_t data);
 
 uint8_t mag_calib_abort(mag *mag_dev);
 
-float mag_read_encoder_angle(mag *mag_dev);
+uint16_t mag_read_encoder_angle(mag *mag_dev);
 
 bool mag_read_device_status(mag *mag_dev);
